[Refactor] Create NewChecklistContextNavigator and inject with bind to SmartChecklistNavigation
