--- features/newchecklist/impl/src/main/kotlin/wottrich/github/io/smartchecklist/newchecklist/navigation/NewChecklistNavigation.kt
@@ -1,35 +1,0 @@
-package wottrich.github.io.smartchecklist.newchecklist.navigation
-
-import androidx.compose.animation.ExperimentalAnimationApi
-import androidx.navigation.NavGraphBuilder
-import androidx.navigation.NavHostController
-import androidx.navigation.navigation
-import wottrich.github.io.smartchecklist.newchecklist.presentation.ui.NewChecklistNameScreen
-import wottrich.github.io.smartchecklist.baseui.navigation.defaultComposableAnimation
-
-@ExperimentalAnimationApi
-fun NavGraphBuilder.newChecklistNavigation(navHostController: NavHostController) {
-    navigation(
-        startDestination = NavigatorNewChecklist.startDestination,
-        route = NavigatorNewChecklist.route
-    ) {
-        defaultComposableAnimation(
-            route = NavigatorNewChecklist.Destinations.NewChecklistName.route
-        ) {
-            NewChecklistNameScreen(
-                onCloseScreen = {
-                    navHostController.popBackStack()
-                }
-            )
-        }
-    }
-}
-
-object NavigatorNewChecklist {
-    val route = "NavigatorNewChecklist"
-    val startDestination = Destinations.NewChecklistName.route
-
-    sealed class Destinations(val route: String) {
-        object NewChecklistName : Destinations("NewChecklistName")
-    }
-}
